+++ src/main/java/swimming/pool/application/swimmingpool/result/SwimmingPoolUpdateResult.java
@@ -1,4 +1,4 @@
-package swimming.pool.application.result;
+package swimming.pool.application.swimmingpool.result;
 
 public class SwimmingPoolUpdateResult {
