+++ src/main/java/swimming/pool/presentation/swimmingpool/DeleteSwimmingPoolController.java
@@ -1,13 +1,11 @@
-package swimming.pool.presentation;
+package swimming.pool.presentation.swimmingpool;
 
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import swimming.pool.application.DeleteSwimmingPoolService;
-import swimming.pool.application.result.SwimmingPoolDeleteResult;
+import swimming.pool.application.swimmingpool.DeleteSwimmingPoolService;
+import swimming.pool.application.swimmingpool.result.SwimmingPoolDeleteResult;
 import swimming.pool.infra.common.ApiResponse;
 
 @RequestMapping("/swimmingpool/delete")
