+++ src/main/java/swimming/pool/presentation/swimmingpool/QuerySwimmingPoolController.java
@@ -1,11 +1,11 @@
-package swimming.pool.presentation;
+package swimming.pool.presentation.swimmingpool;
 
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import swimming.pool.application.QuerySwimmingPoolService;
-import swimming.pool.application.result.SwimmingPoolResult;
+import swimming.pool.application.swimmingpool.QuerySwimmingPoolService;
+import swimming.pool.application.swimmingpool.result.SwimmingPoolResult;
 import swimming.pool.infra.common.ApiResponse;
 
 /*
