+++ src/main/java/swimming/pool/presentation/swimmingpool/UpdateSwimmingPoolController.java
@@ -1,14 +1,14 @@
-package swimming.pool.presentation;
+package swimming.pool.presentation.swimmingpool;
 
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import swimming.pool.application.UpdateSwimmingPoolService;
-import swimming.pool.application.result.SwimmingPoolUpdateResult;
+import swimming.pool.application.swimmingpool.UpdateSwimmingPoolService;
+import swimming.pool.application.swimmingpool.result.SwimmingPoolUpdateResult;
 import swimming.pool.infra.common.ApiResponse;
-import swimming.pool.presentation.request.SwimmingPoolUpdateRequest;
+import swimming.pool.presentation.swimmingpool.request.SwimmingPoolUpdateRequest;
 
 @RestController
 @RequestMapping("/swimmingpool/update")
