+++ src/main/java/swimming/pool/presentation/swimmingpool/request/SwimmingPoolInfoRequest.java
@@ -1,6 +1,6 @@
-package swimming.pool.presentation.request;
+package swimming.pool.presentation.swimmingpool.request;
 
-import swimming.pool.application.command.RegisterSwimmingPoolCommand;
+import swimming.pool.application.swimmingpool.command.RegisterSwimmingPoolCommand;
 public class SwimmingPoolInfoRequest {
 
   private String lastmodts;
